Move Dataset into query API
